requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.0.3
attrs==23.2.0
autopep8==2.0.4
bcrypt==4.1.3
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==24.4.2
blinker==1.7.0
cachetools==5.3.3
cattrs==23.2.3
certifi @ file:///private/tmp/python-certifi-20240201-4765-6ufe0c/certifi-2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer @ file:///private/tmp/python-charset-normalizer-20231123-5225-x00973/charset-normalizer-3.3.2
click==8.1.7
colorgram.py==1.2.0
colorthief==0.2.1
contourpy==1.2.1
cookiecutter==2.6.0
cryptography==42.0.7
customtkinter==5.2.2
cycler==0.12.1
Cython==3.0.10
darkdetect==0.8.0
debugpy==1.8.1
dill==0.3.8
distlib==0.3.8
Django==5.0.6
dnspython==2.6.1
docstring-to-markdown==0.15
docutils==0.21.2
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
flake8==7.0.0
Flask==3.0.2
fleet==0.1.3
flet==0.23.1
flet-core==0.23.1
flet-runtime==0.23.1
fonttools==4.51.0
fuzzywuzzy==0.18.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna @ file:///private/tmp/python-idna-20231125-5279-xtqzsf/idna-3.6
importlib_metadata==7.0.2
inflect==7.0.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
jedi-language-server==0.41.3
Jinja2==3.1.3
joblib==1.4.2
Kivy==2.3.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
Levenshtein==0.25.0
linkify-it-py==2.0.3
lsprotocol==2023.0.1
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mathgenerator==1.5.0
matplotlib==3.8.4
mccabe==0.7.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
mock==5.1.0
MorseCodePy==3.7
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
mypy==1.9.0
mypy-extensions==1.0.0
neovim==0.3.1
nodeenv==1.8.0
npyscreen==4.10.5
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.82
orjson==3.10.5
osascript==2020.12.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
paramiko==3.4.0
parso==0.8.3
pathspec==0.12.1
paypalrestsdk==1.13.3
Pillow==9.5.0
pinscrape==3.2.2
platformdirs==4.2.0
pluggy==1.4.0
prettytable @ git+https://github.com/jazzband/prettytable@f044b0e2c99680c975e5d4523771f819934f4611
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.8
py-wallpaper==0.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
PyBindGen==0.22.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydocstyle==6.3.0
pydotmap==0.1.3
pyflakes==3.2.0
pygame==2.5.2
PyGetWindow==0.0.9
pygls==1.3.0
Pygments==2.18.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pylette==2.2.0
pylint==3.0.4
pylsp-mypy==0.6.8
pymorse==1.4
pyMorseTranslator==1.2.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.7.7
pynvim==0.5.0
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
pyright==1.1.356
PyScreeze==0.1.30
PySocks==1.7.1
PyTermGUI==7.7.1
pytest==7.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.0
python-lsp-black==2.0.0
python-lsp-isort==0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.0
python-multipart==0.0.9
python-slugify==8.0.4
pytoolconfig==1.3.1
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.6.2
repath==0.9.0
requests==2.32.3
rich==13.7.1
rope==1.12.0
rsa==4.9
rubicon-objc==0.4.9
ruff==0.4.3
ruff-lsp==0.0.53
runcmd==2020.12.3
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.19.0
setuptools==70.1.1
shell==1.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
speedtest-cli==2.1.3
sqlparse==0.5.0
starlette==0.37.2
stripe==9.8.0
sympy==1.12.1
tabulate==0.9.0
text-unidecode==1.3
textual==0.71.0
threadpoolctl==3.5.0
tk==0.1.0
tomli==2.0.1
tomlkit==0.12.4
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.9.0
uritemplate==4.1.1
urllib3 @ file:///private/tmp/python-urllib3-20240218-4962-2z5xut/urllib3-2.2.1
urwid==2.6.14
uvicorn==0.30.1
uvloop==0.19.0
validators==0.23.2
virtualenv==20.26.3
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websiteblocker==0.2.6
websockets==12.0
Werkzeug==3.0.1
whatthepatch==1.0.5
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.4/libexec/wheel-0.43.0-py3-none-any.whl#sha256=ff23205a590f2b902a2a21151019e96b93ab4a1409d7c9748b6bd210f4c0f44d
wsproto==1.2.0
yapf==0.40.2
zipp==3.17.0
